Folder, File Open 기능 로직 분리
1. MVVM 패턴 구현
2. 데이터 바인딩을 통한 폴더, 파일 오픈
3. 데이터 바인딩을 통한 Tree, List view
