+++ DefectViewProgram.csproj
@@ -66,6 +66,9 @@
     </Compile>
     <Compile Include="BaseViewModel.cs" />
     <Compile Include="ChipInfo.cs" />
+    <Compile Include="FileOpenViewModel.cs" />
+    <Compile Include="RelayCommadGenerics.cs" />
+    <Compile Include="RelayCommand.cs" />
     <Compile Include="TiffImageLoaderViewModel.cs" />
     <Compile Include="WaferInfo.cs" />
     <Compile Include="Point.cs" />
