+++ MainWindow.xaml
@@ -27,7 +27,7 @@
                        VerticalAlignment="Center"
                        Visibility="Hidden"/>
 
-                    <Image x:Name="CurrentImage" Stretch="Uniform" Visibility="Visible" Source="{Binding CurrentImage}"/>
+                    <!--<Image x:Name="CurrentImage" Stretch="Uniform" Visibility="Visible" Source="{Binding CurrentImage}"/>-->
                 </Grid>
             </Border>
 
@@ -40,14 +40,15 @@
                     </Grid.ColumnDefinitions>
                     <!-- 첫 번째 컬럼: 폴더 트리 -->
                     <DockPanel Grid.Column="0">
-                        <Button DockPanel.Dock="Bottom" Content="폴더선택" Click="OpenFolderButton_Click" Height="30" Margin="5"/>
+                        <Button DockPanel.Dock="Bottom" Content="폴더선택"  Command="{Binding OpenFolderCommand}" Height="30" Margin="5"/>
                         <TreeView x:Name="FolderTreeView" Margin="5" SelectedItemChanged="FolderTreeView_SelectedItemChanged"/>
                     </DockPanel>
 
                     <!-- 두 번째 컬럼: 파일 목록 -->
                     <DockPanel Grid.Column="1">
-                        <Button DockPanel.Dock="Bottom" Content="새로고침" Click="RefreshButton_Click" Height="30" Margin="5"/>
-                        <ListBox x:Name="ListBox" Margin="5" SelectionChanged="ListBox_SelectionChanged" />
+                        <!--<Button DockPanel.Dock="Bottom" Content="새로고침" Click="RefreshButton_Click" Command="{Binding RefreshFileList}" Height="30" Margin="5"/>-->
+                        <ListBox x:Name="FileListBox" Margin="5" ItemsSource="{Binding FileListBox}"/>
+                        <!--SelectionChanged="ListBox_SelectionChanged"-->
                     </DockPanel>
                 </Grid>
             </Border>
@@ -69,50 +70,47 @@
                         <StackPanel Grid.Column="0" Orientation="Vertical" HorizontalAlignment="Center" Margin="35,0,34.6,0">
                             <TextBlock Text="Wafer Information" FontSize="19"/>
                             <ScrollViewer Height="162" VerticalScrollBarVisibility="Auto">
-                                <TextBlock x:Name="waferInfomationText" FontSize="13" TextWrapping="Wrap"/>
+                                <TextBlock x:Name="WaferInfomationText" FontSize="13" TextWrapping="Wrap" Text="{Binding WaferInformation}"/>
                             </ScrollViewer>
                         </StackPanel>
 
-                        <!-- 상단 영역: 디펙 제어 UI -->
+                        <!-- 상단 영역: 디펙 제어 UI --><!--
                         <StackPanel Grid.Column="1" Orientation="Vertical" HorizontalAlignment="Center" Margin="0,0,0.2,0" Width="395">
-                            <!-- 전체 디펙 이동 UI -->
+                            --><!-- 전체 디펙 이동 UI --><!--
                             <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="5">
                                 <Button Content="◀" Width="40" Height="30" Margin="5" Click="PreviousDefectOnWholeWaferButton_Click"/>
-                                <TextBlock x:Name="txtDefectOnWafer" Text="전체 디펙: 0/0" VerticalAlignment="Center" Margin="10,0"/>
+                                <TextBlock x:Name="TextDefectOnWafer" VerticalAlignment="Center" Margin="10,0" Text="{Binding TextDefectOnWafer}"/>
                                 <Button Content="▶" Width="40" Height="30" Margin="5" Click="NextDefectOnWholeWaferButton_Click"/>
                             </StackPanel>
 
-                            <!-- 칩 내부 디펙 이동 UI -->
+                            --><!-- 칩 내부 디펙 이동 UI --><!--
                             <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="5">
                                 <Button Content="◀" Width="40" Height="30" Margin="5" Click="PreviousDefectOnChipButton_Click"/>
                                 <TextBlock x:Name="txtDefectOnChip" Text="칩 내 디펙: 0/0" VerticalAlignment="Center" Margin="10,0"/>
                                 <Button Content="▶" Width="40" Height="30" Margin="5" Click="NextDefectOnChipButton_Click"/>
                             </StackPanel>
 
-                            <!-- 칩 디펙 조회 입력 -->
+                            --><!-- 칩 디펙 조회 입력 --><!--
                             <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="5">
                                 <TextBlock Text="칩 좌표:" VerticalAlignment="Center" Margin="5,0"/>
                                 <TextBox x:Name="transferCoordinate1" Width="80" VerticalAlignment="Center" Margin="5,0"/>
                                 <Button Content="칩 Defect 조회" Width="100" Height="30" Margin="5" Click="TransferCoordinateButton_Click"/>
                                 <Button Content="전체 Defect 조회" Width="100" Height="30" Margin="5" Click="ShowAllList"/>
                             </StackPanel>
-                        </StackPanel>
+                        </StackPanel>-->
                     </Grid>
-                                       
-                    
-                    
-                    
+                                                  
                     <!-- 하단 영역: 데이터그리드 -->
                     <DockPanel Grid.Row="1" LastChildFill="True">
-                        <DataGrid x:Name="defectList" AutoGenerateColumns="False">
+                        <DataGrid x:Name="DefectList" AutoGenerateColumns="False" ItemsSource="{Binding DefectList}">
                             <DataGrid.Columns>
-                                <DataGridTextColumn Header="ID" Binding="{Binding Path=id}" Width="50"/>
-                                <DataGridTextColumn Header="XREL" Binding="{Binding Path=xrel}" Width="100"/>
-                                <DataGridTextColumn Header="YREL" Binding="{Binding Path=yrel}" Width="100"/>
-                                <DataGridTextColumn Header="XIndex" Binding="{Binding Path=xindex}" Width="50"/>
-                                <DataGridTextColumn Header="YIndex" Binding="{Binding Path=yindex}" Width="50"/>
-                                <DataGridTextColumn Header="XSize" Binding="{Binding Path=xsize}" Width="50"/>
-                                <DataGridTextColumn Header="YSize" Binding="{Binding Path=ysize}" Width="50"/>
+                                <DataGridTextColumn Header="ID" Binding="{Binding Path=DefectId}" Width="50"/>
+                                <DataGridTextColumn Header="XREL" Binding="{Binding Path=XRel}" Width="100"/>
+                                <DataGridTextColumn Header="YREL" Binding="{Binding Path=YRel}" Width="100"/>
+                                <DataGridTextColumn Header="XIndex" Binding="{Binding Path=XIndex}" Width="50"/>
+                                <DataGridTextColumn Header="YIndex" Binding="{Binding Path=YIndex}" Width="50"/>
+                                <DataGridTextColumn Header="XSize" Binding="{Binding Path=XSize}" Width="50"/>
+                                <DataGridTextColumn Header="YSize" Binding="{Binding Path=YSize}" Width="50"/>
                             </DataGrid.Columns>
                         </DataGrid>
                     </DockPanel>
