+++ MainWindow.xaml.cs
@@ -14,428 +14,32 @@ namespace DefectViewProgram
     /// </summary>
     public partial class MainWindow : Window
     {
+
+        private FileOpenViewModel fileOpenViewModel;
         public MainWindow()
         {
             InitializeComponent();
             DataContext = new TiffImageLoaderViewModel();
-        }
+            DataContext = new FileOpenViewModel();
 
-        private string currentFolderPath;
-        public string CurrentFolderPath
-        {
-            get => currentFolderPath;
-            set => currentFolderPath = value;
-        }
-        public string FullPath { get; set; }
+            // 뷰모델 인스턴스 가져오기
+            fileOpenViewModel = DataContext as FileOpenViewModel;
 
-        public bool isChipDataView = false;
-
-
-        private string selectedFileName;
-        public string SelectedFileName
-        {
-            get => selectedFileName;
-            set => selectedFileName = value;
-        }
-
-        // 전체 디펙개수
-        public int wholeWaferDefectCount = 0;
-
-        // 전체 칩 인덱스
-        public int currentWaferIndex = 0;
-
-        // 칩 내부 디펙 현재 인덱스
-        public int currentChipDefectIndex = 0;
-
-        // 선택된 Tiff 이미지 인덱스
-        public int currentImageIndex = 0;
-
-        private ObservableCollection<object> allDefectsItems = new ObservableCollection<object>();
-        private ObservableCollection<object> chipDefectsItems = new ObservableCollection<object>();
-
-
-        public void OpenFolderButton_Click(object sender, RoutedEventArgs e)
-        {
-            var dialog = new FolderBrowserDialog();
-            if (dialog.ShowDialog() == System.Windows.Forms.DialogResult.OK)
+            // UI 요소 연결
+            if (fileOpenViewModel != null)
             {
-                // 트리뷰에 폴더 추가
-                LoadFolders(dialog.SelectedPath);
+                fileOpenViewModel.FolderTreeView = this.FolderTreeView;
+                fileOpenViewModel.FileListBox = this.FileListBox;
             }
         }
 
-        public void LoadFolders(string path) // 탐색기에서 폴더 선택했을때
+        // 트리뷰 선택 변경 이벤트 핸들러
+        private void FolderTreeView_SelectedItemChanged(object sender, RoutedPropertyChangedEventArgs<object> e)
         {
-            FolderTreeView.Items.Clear();
-            CurrentFolderPath = path;
-
-            try
+            if (fileOpenViewModel != null)
             {
-                var rootDir = new DirectoryInfo(path);
-                var rootItem = CreateTreeItem(rootDir);
-                FolderTreeView.Items.Add(rootItem);
-                rootItem.IsExpanded = true;
+                fileOpenViewModel.FolderTreeView_SelectedItemChanged(sender, e);
             }
-            catch (Exception ex)
-            {
-            }
-        }
-
-        // 탐색기에서 폴더 선택했을때
-        public TreeViewItem CreateTreeItem(DirectoryInfo dirInfo)
-        {
-            var item = new TreeViewItem { Header = dirInfo.Name, Tag = dirInfo.FullName };
-
-            try
-            {
-                foreach (var subDir in dirInfo.GetDirectories())
-                {
-                    // 시스템/숨김 폴더 제외
-                    if (!subDir.Attributes.HasFlag(FileAttributes.Hidden) &&
-                        !subDir.Attributes.HasFlag(FileAttributes.System))
-                    {
-                        item.Items.Add(CreateTreeItem(subDir));
-                    }
-                }
-            }
-            catch { /* 접근 권한 없는 폴더 무시 */ }
-
-            return item;
-        }
-
-
-        // 탐색기에서 폴더 선택했을때, 혹은 내부 트리뷰에서 폴더 선택했을때 
-        public void FolderTreeView_SelectedItemChanged(object sender, RoutedPropertyChangedEventArgs<object> e)
-        {
-            var item = FolderTreeView.SelectedItem as TreeViewItem;
-            if (item != null)
-            {
-                LoadFiles(item.Tag.ToString());
-            }
-        }
-
-        public void LoadFiles(string path)
-        {
-            ListBox.Items.Clear();
- 
-                CurrentFolderPath = path;
-                foreach (var file in Directory.GetFiles(path))
-                {
-                    if (!IsKlarfInfoFileCheck(file))
-                    {
-                        continue;
-                    }
-                    ListBox.Items.Add(Path.GetFileName(file));
-                }
-        }
-
-        public bool IsKlarfInfoFileCheck(string filePath)
-        {
-            try
-            {
-                using (StreamReader reader = new StreamReader(filePath))
-                {
-                    // 처음 20줄만 확인
-                    for (int i = 0; i < 20 && !reader.EndOfStream; i++)
-                    {
-                        string line = reader.ReadLine();
-                        if (line.Contains("FileVersion") || line.Contains("WaferID") || line.Contains("LotID"))
-                        {
-                            return true;
-                        }
-                    }
-                }
-                return false;
-            }
-            catch
-            {
-                return false; // 파일 읽기 실패시
-            }
-        }
-
-        /// <summary>
-        /// 파일 새로고침
-        /// </summary>
-        public void RefreshButton_Click(object sender, RoutedEventArgs e)
-        {
-            var item = FolderTreeView.SelectedItem as TreeViewItem;
-            if (item != null)
-            {
-                LoadFiles(item.Tag.ToString());        
-            }
-            ChipInfo chip = new ChipInfo();
-            chip.ChipDefectClear();
-        }
-
-        /// <summary>
-        /// 폴더 Path 가져오는 함수
-        /// </summary>
-        /// 
-        public string GetSelectedFolderPath()
-        {
-            using (var folderBrowswerDialog = new FolderBrowserDialog())
-            {
-                DialogResult result = folderBrowswerDialog.ShowDialog();
-
-                if(result == System.Windows.Forms.DialogResult.OK)
-                {
-                    return folderBrowswerDialog.SelectedPath;
-                }
-                else
-                {
-                    return null;
-                }
-            }
-        }
-
-
-        public void ShowAllList(object sender, RoutedEventArgs e)
-        {
-            defectList.ItemsSource = allDefectsItems;
-            currentWaferIndex = 0;
-            currentChipDefectIndex = 0;
-            txtDefectOnWafer.Text = $"전체 디펙: 1/{allDefectsItems.Count}";
-
-            //TiffImageLoaderViewModel loader = new TiffImageLoaderViewModel();
-            var vm = this.DataContext as TiffImageLoaderViewModel;
-            vm.LoadTiffImage(CurrentFolderPath);
-            isChipDataView = false;
-        }
-
-        /// <summary>
-        /// 리스트 박스에 있는것(Klarf.txt) 클릭했을 때 파싱해주는 함수
-        /// </summary>
-        /// 
-        public void ListBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            if (ListBox.SelectedItem != null)
-            {
-                ChipInfo chip = new ChipInfo();
-                chip.ChipDefectClear();
-
-                SelectedFileName = ListBox.SelectedItem.ToString();
-                FullPath = Path.Combine(CurrentFolderPath, SelectedFileName);
-
-                Console.WriteLine(FullPath);
-
-                KlarfFileParser parser = new KlarfFileParser();
-               
-                parser.ParseText(FullPath);
-
-                string defectInfo = chip.GetAllDefects();
-
-                //ObservableCollection<object> items = new ObservableCollection<object>();
-
-                string[] lines = defectInfo.Split('\n');
-
-                allDefectsItems.Clear();
-
-                foreach (string line in lines)
-                {
-                    string[] parts = line.Split(',');
-
-                    if (parts.Length < 7) continue;
-
-                    // 익명 객체 사용
-                    allDefectsItems.Add(new
-                    {
-                        id = parts[2],
-                        xrel = parts[3],
-                        yrel = parts[4],
-                        xindex = parts[5],
-                        yindex = parts[6],
-                        xsize = parts[7],
-                        ysize = parts[8]
-                    });
-                }
-
-                waferInfomationText.Text = parser.waferInfo;
-
-                defectList.ItemsSource = allDefectsItems;
-                currentWaferIndex = 0;
-                currentChipDefectIndex = 0;
-                txtDefectOnWafer.Text = $"전체 디펙: 1/{allDefectsItems.Count}";
-
-                var vm = this.DataContext as TiffImageLoaderViewModel;
-                vm.LoadTiffImage(CurrentFolderPath);
-
-                //Console.WriteLine(items);
-            }
-        }
-
-
-        /// <summary>
-        /// 버튼 눌렀을때 특정 칩의 디펙을 보여주는 함수
-        /// </summary>
-        /// 
-        public void TransferCoordinateButton_Click(object sender, RoutedEventArgs e)
-        {
-            string temp = transferCoordinate1.Text;
-
-            // 웨이퍼 그래픽에서 입력 됐을때 들어오는 속성추가해야됨 04.02
-
-            string[] tempArr = temp.Split(',');
-
-            for (int i = 0; i < tempArr.Length; i++)
-            {
-                tempArr[i] = tempArr[i].Trim();
-            }
-
-            int x = int.Parse(tempArr[0]);
-            int y = int.Parse(tempArr[1]);
-           
-            ChipInfo chip = new ChipInfo();
-           
-            //ObservableCollection<object> items = new ObservableCollection<object>();
-
-            List<DefectInfo> defectInfo = chip.GetDefects(x, y);
-
-            if(defectInfo.Count == 0)
-            {
-                txtDefectOnChip.Text = $"칩 내 디펙: {0}/{0}";
-                ClearDefectImage();
-                return;
-            }
-
-            currentImageIndex = 0;
-
-            chipDefectsItems.Clear();
-
-            for (int i = 0; i < defectInfo.Count; i++)
-            {
-                chipDefectsItems.Add(new
-                {
-                    id = defectInfo[i].DefectId,
-                    xrel = defectInfo[i].XRel,
-                    yrel = defectInfo[i].YRel,
-                    xindex = defectInfo[i].XIndex,
-                    yindex = defectInfo[i].YIndex,
-                    xsize = defectInfo[i].XSize,
-                    ysize = defectInfo[i].YSize
-                });
-            }
-
-            // 칩 모드로 전환
-            // 먼저 ItemsSource 설정
-            defectList.ItemsSource = chipDefectsItems;
-
-            // 아이템이 있는지 확인
-            if (chipDefectsItems.Count > 0)
-            {
-                // 첫 아이템 선택
-                defectList.SelectedIndex = 0;
-                currentChipDefectIndex = 0;
-
-                // 이제 SelectedItem 접근
-                dynamic selectedItem = defectList.SelectedItem;
-                currentImageIndex = selectedItem != null ? (int)selectedItem.id - 1 : -2;
-            }
-            isChipDataView = true;
-            txtDefectOnChip.Text = $"칩 내 디펙: {currentChipDefectIndex + 1}/{chipDefectsItems.Count}";
-            var vm = this.DataContext as TiffImageLoaderViewModel;
-            vm.LoadDefectImageFromChipOnSelected(currentImageIndex);
-        }
-
-
-
-        public void NextDefectOnWholeWaferButton_Click(object sender, RoutedEventArgs e)
-        {
-            if (isChipDataView)
-            {
-                return;
-            }
-
-            currentWaferIndex++;
-            int count = defectList.Items.Count;
-
-            defectList.SelectedIndex = currentWaferIndex;
-
-            txtDefectOnWafer.Text =  $"전체 디펙: {currentWaferIndex + 1}/{count}";
-            var vm = this.DataContext as TiffImageLoaderViewModel;
-            vm.LoadDefectImageFromWholeSelected(currentWaferIndex);
-        }
-
-
-        public void PreviousDefectOnWholeWaferButton_Click(object sender, RoutedEventArgs e)
-        {
-
-            if (isChipDataView)
-            {
-                return;
-            }
-
-            if (currentWaferIndex > 0)
-            {
-                currentWaferIndex--;
-                wholeWaferDefectCount = defectList.Items.Count;
-
-                defectList.SelectedIndex = currentWaferIndex;
-
-                txtDefectOnWafer.Text = $"전체 디펙: {currentWaferIndex + 1}/{wholeWaferDefectCount}";
-                var vm = this.DataContext as TiffImageLoaderViewModel;
-                vm.LoadDefectImageFromWholeSelected(currentWaferIndex);
-            }
-        }
-
-
-        // 칩 내부 다음 결함
-        public void NextDefectOnChipButton_Click(object sender, RoutedEventArgs e)
-        {
-            if (!isChipDataView)
-            {
-                return;
-            }
-
-            if (chipDefectsItems.Count == 0) return;
-
-            if (currentChipDefectIndex < chipDefectsItems.Count - 1)
-            {
-                currentChipDefectIndex++; // 리스트 기준 인덱스 증가
-                defectList.SelectedIndex = currentChipDefectIndex; // 리스트에서 해당 인덱스 선택
-
-                // 선택된 아이템에서 defectId 가져오기
-                dynamic selectedItem = defectList.SelectedItem;
-                currentImageIndex = selectedItem != null ? (int)selectedItem.id -1: 0;
-
-                txtDefectOnChip.Text = $"칩 내 디펙: {currentChipDefectIndex + 1}/{chipDefectsItems.Count}";
-
-                // 올바른 id로 이미지 로드
-                var vm = this.DataContext as TiffImageLoaderViewModel;
-                vm.LoadDefectImageFromChipOnSelected(currentImageIndex);
-            }
-        }
-
-
-        // 칩 내부 이전 결함
-        public void PreviousDefectOnChipButton_Click(object sender, RoutedEventArgs e)
-        {
-            if (!isChipDataView)
-            {
-                return;
-            }
-
-            if (chipDefectsItems.Count == 0) return;
-
-            if (currentChipDefectIndex > 0)
-            {
-                currentChipDefectIndex--;
-                defectList.SelectedIndex = currentChipDefectIndex;
-
-                // 선택된 아이템에서 defectId 가져오기
-                dynamic selectedItem = defectList.SelectedItem;
-                currentImageIndex = selectedItem != null ? (int)selectedItem.id - 1 : 0;
-
-                txtDefectOnChip.Text = $"칩 내 디펙: {currentChipDefectIndex + 1}/{chipDefectsItems.Count}";
-
-                var vm = this.DataContext as TiffImageLoaderViewModel;
-                vm.LoadDefectImageFromChipOnSelected(currentImageIndex);
-
-            }
-        }
-        public void ClearDefectImage()
-        {
-            CurrentImage.Source = null;
-            NoImageText.Visibility = Visibility.Visible;
         }
     }
 }
